requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.0
django-crispy-forms==1.11.2
django-tenant-schemas==1.10.0
django-tenants==3.2.1
ordered-set==4.0.2
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
wincertstore==0.2
